requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
Django==4.2.1
django-appconf==1.0.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-modeltranslation==0.18.9
django-redis==5.2.0
django-rest-swagger==2.2.0
django-staticfiles==1.2.1
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
frozenlist==1.3.3
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
multidict==6.0.4
openapi-codec==1.3.2
prompt-toolkit==3.0.38
PyJWT==2.7.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
requests==2.30.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
twilio==8.2.1
typing_extensions==4.6.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
